Feature/life improvements
looks good. Thanks for the updates!

+++ provisions/docker/rebuild-and-collect-static-files.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if (( EUID != 0 )); then
+   echo "Script must be run as root."
+   exit 126
+fi
+
 (set -x; docker run --rm -t \
     --volumes-from app \
     datahuborg/datahub \
